+++ lib/widgets/title.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+Widget titleWidget(String s) {
+  return Text(
+    s,
+    style: const TextStyle(
+        color: Colors.black, fontWeight: FontWeight.bold, fontSize: 20.0),
+  );
+}
